requirements.txt
openai==1.55.3
google-cloud-bigquery==3.25.0
streamlit==1.36.0
streamlit-toggle-switch==1.0.2